requirements.txt
google-generativeai>=0.3.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
PyYAML>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pypdf>=3.0.0
